Fix trading

+++ game/src/main/kotlin/world/gregs/voidps/world/community/trade/lend/Loan.kt
@@ -55,7 +55,7 @@ object Loan {
     fun getTimeRemaining(player: Player, timeKey: String): Long {
         return when {
             player.contains(timeKey) -> {
-                val timeout: Int = player.getOrNull(timeKey) ?: return -1
+                val timeout: Long = player.getOrNull(timeKey) ?: return -1
                 System.currentTimeMillis() - timeout
             }
             else -> 0
